requirements.txt
streamlit>=1.29.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-cohere>=0.1.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pymupdf>=1.23.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=10.0.0
typing-extensions>=4.5.0
pytesseract>=0.3.10
cohere>=5.0.0
langchain-groq>=0.1.0